pyproject.toml
[project]
name = "orchestrai"
version = "0.1.0"
description = ""
authors = [
    {name = "Akash Das",email = "akashdas1966@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai (>=1.98.0,<2.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "google-genai (>=1.28.0,<2.0.0)",
    "faiss-cpu (>=1.11.0.post1,<2.0.0)"
]

[tool.poetry]
packages = [{include = "orchestrai", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
